tgbot/pyproject.toml
[project]
name = "tgbot"
version = "0.1.0"
description = "Telegram bot for AI Daily Digest - Sends daily summaries and provides chat functionality"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "supabase>=2.17.0",
    "apscheduler>=3.11.0",
    "aiohttp>=3.12.15",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "openai>=1.98.0",
    "python-telegram-bot>=22.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tgbot"]
